backend/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
bcrypt
passlib
python-dotenv
python-jose
python-multipart